backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pandas
cachetools
pydantic-settings
python-dotenv
email-validator
openpyxl
python-multipart